--- packages/frontend/component/src/ui/lottie/collections-icon.css.ts
@@ -1,9 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const root = style({
-  width: '1em',
-  height: '1em',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-});
